scorer/requirements.txt
groq==0.11.0
supabase==2.4.2
python-dotenv==1.0.1
